# community-backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ServiceResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ServiceResponse_any-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_any-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ProjectCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"draft",
							"in_progress",
							"completed",
							"on_hold",
							"cancelled"
						]
					},
					"targetGroup": {
						"type": "string",
						"nullable": true
					},
					"projectDuration": {
						"type": "string",
						"nullable": true
					},
					"geographicArea": {
						"type": "string",
						"nullable": true
					},
					"stakeholderIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"donorIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"documents": {
						"items": {
							"properties": {
								"deleteToken": {
									"type": "string",
									"nullable": true
								},
								"publicId": {
									"type": "string",
									"nullable": true
								},
								"userId": {
									"type": "string"
								},
								"documentUrl": {
									"type": "string",
									"nullable": true
								},
								"addedAt": {
									"type": "string",
									"format": "date-time"
								},
								"type": {
									"type": "string",
									"nullable": true
								},
								"size": {
									"type": "number",
									"format": "double",
									"nullable": true
								},
								"documentName": {
									"type": "string"
								}
							},
							"required": [
								"userId",
								"documentName"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectUpdateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"draft",
							"in_progress",
							"completed",
							"on_hold",
							"cancelled"
						]
					},
					"targetGroup": {
						"type": "string"
					},
					"projectDuration": {
						"type": "string"
					},
					"geographicArea": {
						"type": "string"
					},
					"stakeholderIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"donorIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"documents": {
						"items": {
							"properties": {
								"deleteToken": {
									"type": "string"
								},
								"publicId": {
									"type": "string"
								},
								"userId": {
									"type": "string"
								},
								"documentUrl": {
									"type": "string"
								},
								"addedAt": {
									"type": "string",
									"format": "date-time"
								},
								"type": {
									"type": "string"
								},
								"size": {
									"type": "number",
									"format": "double"
								},
								"documentName": {
									"type": "string"
								}
							},
							"required": [
								"userId",
								"documentName"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"PermissionCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PermissionUpdateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"DocumentCreateRequest": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"documentName": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"type": {
						"type": "string",
						"nullable": true
					},
					"addedAt": {
						"type": "string",
						"format": "date-time"
					},
					"documentUrl": {
						"type": "string",
						"nullable": true
					},
					"userId": {
						"type": "string"
					},
					"publicId": {
						"type": "string",
						"nullable": true
					},
					"deleteToken": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"documentName",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "community-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Bikorimana Saveur"
		}
	},
	"paths": {
...
		"/api/roles/{roleId}/permissions": {
			"post": {
				"operationId": "AddPermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"role:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"permissionId": {
										"type": "string"
									}
								},
								"required": [
									"permissionId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/roles/{roleId}/permissions/{permissionId}": {
			"delete": {
				"operationId": "RemovePermission",
				"responses": {
					"204": {
						"description": "No Content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"role:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "permissionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/projects": {
			"get": {
				"operationId": "GetProjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": [
							"project:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": [
							"project:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/projects/{projectId}": {
			"get": {
				"operationId": "GetProjectById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Project not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": [
							"project:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": [
							"project:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProject",
				"responses": {
					"204": {
						"description": "No Content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": [
							"project:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/permissions": {
			"get": {
				"operationId": "ListPermissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Permissions"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Permissions"
				],
				"security": [
					{
						"jwt": [
							"permission:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PermissionCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/permissions/{permissionId}": {
			"get": {
				"operationId": "GetPermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Permissions"
				],
				"security": [
					{
						"jwt": [
							"permission:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "permissionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Permissions"
				],
				"security": [
					{
						"jwt": [
							"permission:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "permissionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PermissionUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Permissions"
				],
				"security": [
					{
						"jwt": [
							"permission:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "permissionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/documents": {
			"get": {
				"operationId": "GetDocuments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "projectId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateDocument",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DocumentCreateRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}